requirements.txt
flask==2.2.3
flask-cors==3.0.10
google-cloud-bigquery==3.11.0
google-cloud-storage==2.10.0
gunicorn==20.1.0
